Edited stuff
Edited the location of the Meet My Characters and Stories nav to before the Contact Me tab and I also changed the color of the text in all of the locations.

I also removed the footer from the stylesheet, I removed the background url duplicate, and I got rid of body * because I don't think that I need it.

I also updated the Read Me in the root folder and the rest of the .html's except for Blog 2.html and 404.html with temporary copyrights until I can afford a proper copyright.

I also updated Fantasy.html and Hub.html with a disclaimer regarding not owning the copyright to Disney's Toontown Online and that I don't own the stuff yada-yada-yada basic boring stuff to not get sued by Disney and/or a Karen.

I also updated TODO.txt to fit with my current plan.

That's it!