Edited stuff
I moved the 1st blog post into a subfolder titled 1 and I changed the location of it on Blog.html

I essentially did the same with the Tranmissions where I put the Transmission into a subfolder titled 1 and changed the URL for it as well!

I also renamed the Blogs folder to just Blog.