Edited stuff
This is the second and hopefully final patch for the navigation bar which I messed up.

On My_Characters.html, I changed the title and I fixed the stylesheet location there.

I added the nav bar to the Error 404 page.

I fixed the closing paragraph in The Hub.html where it had said </o> before, it now says </p>