Edited everything
I've added a white-colored <hr> tag to the top and bottom's of every .html listed here.

I changed the color of the About pngs to Lilac Purple which is Destiny's new dialog color of choice.

I also changed the color of Destiny's dialog to Lilac Purple as stated above.

I centered the text on the Blog just for now until I get a second opinion on what looks better.